ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Jett Chen", email = "jettchen12345@gmail.com" }]
dependencies = [
    "geoclip>=1.2.0",
    "pandas>=2.2.2",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "geojson>=3.1.0",
    "ruff>=0.4.0",
    "shapely>=2.0.4",
    "ipykernel>=6.29.4",
    "pyproject-local-kernel>=0.5.5",
    "python-dotenv>=1.0.1",
    "geopy>=2.4.1",
    "modal>=0.62.98",
    "pytorch-lightning>=2.2.3",
    "pytorch-metric-learning>=2.5.0",
    "torchmetrics>=1.0.3",
    "timm>=0.9.16",
    "matplotlib>=3.8.4",
    "fastapi[standard]>=0.110.3",
    "pydantic>=2.7.1",
    "python-multipart>=0.0.9",
    "pyequilib",
    "opentelemetry-distro>=0.45b0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "openai>=1.30.4",
    "pinecone-client>=4.1.0",
    "httpx[socks]>=0.27.0",
    "httpcore[socks]>=1.0.5",
    "pandarallel>=1.6.5",
    "websockets>=12.0",
    "cattrs==23.2.3",
    "pyjwt[crypto]>=2.8.0",
    "supabase>=2.5.1",
    "redis>=5.0.6",
    "sentry-sdk[fastapi]>=2.7.1",
    "requests>=2.32.3",
    "upstash-ratelimit>=1.1.0",
    "upstash-redis>=1.1.0",
    "ek-geo>=0.2.1",
    "replicate>=0.32.1",
    "tenacity>=9.0.0",
    "filetype>=1.2.0",
    "upstash-vector>=0.5.0",
    "numpy",
    "opencv-python",
    "Pillow",
    "tqdm>=4.36.0",
    "plotly",
    "scipy",
    "omegaconf",
    "lxml",
    "rtree",
    "scikit-learn",
    "exifread",
    "urllib3>=2",
    "aiostream>=0.7.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "memray>=1.13.4",
    "py-spy>=0.3.14",
    "perspective2d @ git+https://github.com/jinlinyi/PerspectiveFields.git",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ml"]
